refactor(shared): name folder service
