+++ src/shared/components/Drawer/components/Aside/components/DialogNewBoard/services/fetchPostBoardService.ts
@@ -1,6 +1,6 @@
 import { IFetchResponseDefault } from "shared/common/types/Fetch";
 import api from "shared/services/api";
-import { IFetchPostBoardService } from "shared/components/Drawer/components/Aside/components/DialogNewBoard/service/types";
+import { IFetchPostBoardService } from "shared/components/Drawer/components/Aside/components/DialogNewBoard/services/types";
 
 export const fetchPostBoardService: IFetchPostBoardService = async ({
 	body,
