+++ src/shared/components/Drawer/components/Aside/components/DialogNewBoard/services/types/fetchPostBoardService.types.ts
